[refactor] 전체 조회 / mapper 함수 패키지 이동 (#376)
