+++ android/app/src/main/java/com/teamtripdraw/android/data/model/mapper/PostDataMapper.kt
@@ -1,10 +1,5 @@
 package com.teamtripdraw.android.data.model.mapper
 
-import com.teamtripdraw.android.data.httpClient.dto.response.AddPostResponse
-import com.teamtripdraw.android.data.httpClient.dto.response.GetPostPointResponse
-import com.teamtripdraw.android.data.httpClient.dto.response.GetTripPostListResponse
-import com.teamtripdraw.android.data.httpClient.dto.response.GetTripPostResponse
-import com.teamtripdraw.android.data.model.DataPoint
 import com.teamtripdraw.android.data.model.DataPost
 import com.teamtripdraw.android.data.model.DataPostOfAll
 import com.teamtripdraw.android.data.model.DataPrePatchPost
@@ -14,36 +9,6 @@ import com.teamtripdraw.android.domain.model.post.PostOfAll
 import com.teamtripdraw.android.domain.model.post.PrePatchPost
 import com.teamtripdraw.android.domain.model.post.PrePost
 import java.time.LocalDateTime
-
-fun AddPostResponse.toData(): Long {
-    return postId
-}
-
-fun GetPostPointResponse.toData(): DataPoint {
-    return DataPoint(
-        pointId = pointId,
-        latitude = latitude,
-        longitude = longitude,
-        hasPost = hasPost,
-        recordedAt = recordedAt,
-    )
-}
-
-fun GetTripPostResponse.toData(): DataPost {
-    return DataPost(
-        postId = postId,
-        tripId = tripId,
-        title = title,
-        writing = writing,
-        address = address,
-        point = point.toData(),
-        postImageUrl = postImageUrl,
-        routeImageUrl = routeImageUrl,
-    )
-}
-
-fun GetTripPostListResponse.toData(): List<DataPost> =
-    this.posts.map { it.toData() }
 
 fun PrePost.toData(): DataPrePost {
     return DataPrePost(
